alfred: after_quiet, an alternative to rate_limit that resets the timer
rate_limit will run a function no more often than the specified value,
and run it at that time from the first time it was requested.
after_quiet will run a function no more often than the specified value,
and run it after there are been no requests for it for that time.

Quirks:
- after_quiet will have the time-extending effect on functions that
  were registered with rate_limit originally
- each call to after_quiet will re-calculate the time to defer until,
  even if that time is before an earlier call would have had it run.