Change apteryx-sync destination configuration path
This was previously configured as /apteryx/sync/dest_a/* = URL.
Using a * for this field does not make much sense.

This has been changed to /apteryx-sync/destinations/dest_a = URL.

In addition, the path is based on new apteryx_sync.h header file, rather
than apteryx.h, to complete the separation of the two.

Signed-off-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>