Fix generation of paths of synced variables
Paths should be generated from the parent of the leaf node in
the tree, not from the leaf and then stripping off the last '/'.
Previous method fails when value include '/' characters.

In one instance the excluded path list was being checked against
entire path/value concatenation which then failed to exclude.

Remove an unused global variable.